Separate file for exponential formatting.
