+++ PowerT/Controls/MainWindow.cs
@@ -6,7 +6,6 @@ using PowerT.Data;
 using PowerT.Properties;
 using System.Net;
 using System.Text;
-using System.Text.RegularExpressions;
 using System.Windows.Forms.DataVisualization.Charting;
 
 namespace PowerT.Controls;
@@ -14,9 +13,6 @@ namespace PowerT.Controls;
 [DesignerCategory("Code")]
 internal sealed partial class MainWindow : Form
 {
-    [GeneratedRegex(@"(?<mantissa>.*)(E(?<exponent>.*))")]
-    private static partial Regex re_expFormat();
-
     private readonly TextBox tb_sources;
     private readonly SplitContainer _container, _params_container;
     private readonly Chart _chart;
@@ -202,7 +198,7 @@ internal sealed partial class MainWindow : Form
             Minimum = 0.001M,
             Maximum = 1_000M,
             Value = (decimal)this.axisX.Minimum,
-            Formatter = ExpFormatter,
+            Formatter = UIUtils.ExpFormatter,
             Parent = this._params_container.Panel2,
         };
         this.nud_timeFrom.ValueChanged += (sender, e) => this.axisX.Minimum = (double)this.nud_timeFrom.Value;
@@ -223,7 +219,7 @@ internal sealed partial class MainWindow : Form
             Minimum = 5M,
             Maximum = 1_000_000M,
             Value = (decimal)this.axisX.Maximum,
-            Formatter = ExpFormatter,
+            Formatter = UIUtils.ExpFormatter,
             Parent = this._params_container.Panel2,
         };
         this.nud_timeTo.ValueChanged += (sender, e) => this.axisX.Maximum = (double)this.nud_timeTo.Value;
@@ -252,7 +248,7 @@ internal sealed partial class MainWindow : Form
             Minimum = 0.001M,
             Maximum = 1_000M,
             Value = (decimal)this.axisY.Minimum,
-            Formatter = ExpFormatter,
+            Formatter = UIUtils.ExpFormatter,
             Parent = this._params_container.Panel2,
         };
         this.nud_signalFrom.ValueChanged += (sender, e) => this.axisY.Minimum = (double)this.nud_signalFrom.Value;
@@ -273,7 +269,7 @@ internal sealed partial class MainWindow : Form
             Minimum = 50M,
             Maximum = 1_000_000M,
             Value = (decimal)this.axisY.Maximum,
-            Formatter = ExpFormatter,
+            Formatter = UIUtils.ExpFormatter,
             Parent = this._params_container.Panel2,
         };
         this.nud_signalTo.ValueChanged += (sender, e) => this.axisY.Maximum = (double)this.nud_signalTo.Value;
@@ -520,48 +516,6 @@ internal sealed partial class MainWindow : Form
             this._container.Height = this.Height - 130;
         };
     } // ctor ()
-
-    private static string ExpFormatter(decimal value)
-    {
-        var s = value.ToString("E2");
-
-        var match = re_expFormat().Match(s);
-        if (!match.Success) return s;
-
-        var mantissa = match.Groups["mantissa"].Value;
-        var exponent = match.Groups["exponent"].Value;
-
-        var sb = new StringBuilder(mantissa);
-        sb.Append("×10");
-        if (exponent.StartsWith('-'))
-            sb.Append('⁻');  // U+207B
-
-        var e = exponent[1..].TrimStart('0');
-        if (e.Length == 0)
-        {
-            sb.Append('⁰');  // U+2070
-        }
-        else
-        {
-            // append unicode superscript
-            foreach (var c in exponent[1..].TrimStart('0'))
-            {
-                /*
-                 * Superscript of 1, 2, 3 are located in U+00Bx,
-                 * whereas the rest in U+207x.
-                 */
-                sb.Append(c switch
-                {
-                    '1' => '¹',  // U+00B9
-                    '2' => '²',  // U+00B2
-                    '3' => '³',  // U+00B3
-                    _ => (char)(c + 0x2040)  // U+2070 - U+2079
-                });
-            }
-        }
-
-        return sb.ToString();
-    } // private static string ExpFormatter (decimal)
 
     private void SourcesDragEnter(object? sender, DragEventArgs e)
     {
